app/backend/requirements.txt
crewai
crewai_tools
langchain_community  
langchain_openai 
qdrant-client 
pdfplumber 
langchainhub
python-dotenv
uvicorn
fastapi
streamlit
pydantic
redis
dnspython
llama-index 
llama-index-graph-stores-neo4j
langgraph